+++ src/utils/types.js
@@ -1,0 +1,13 @@
+import PropTypes from 'prop-types';
+
+const ingredientType = {
+    data: PropTypes.arrayOf(PropTypes.shape({
+        _id: PropTypes.string.isRequired,
+        type: PropTypes.oneOf(['bun', 'sauce', 'main']).isRequired,
+        name: PropTypes.string.isRequired,
+        image: PropTypes.string.isRequired,
+        price: PropTypes.number.isRequired,
+    })).isRequired,
+};
+
+export { ingredientType };
